pyproject.toml
[project]
name = "muf-up"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "ruamel-yaml>=0.19.1",
]
